backend/requirements.txt
fastapi[standard]
langchain
langchain-openai
pydantic
httpx
python-dotenv
instructor
langgraph
